pyproject.toml
[project]
name = "signal"
version = "0.1.0"
description = "AI-digested language + LLM research, delivered beautifully"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "feedparser>=6.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
signal = "signalfeed.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/signalfeed"]
